Split transcript response extraction
