+++ FoundationLabCore/Sources/FoundationLabCore/Providers/Transcript+LatestResponse.swift
@@ -1,0 +1,17 @@
+import FoundationModels
+
+extension Transcript {
+    func latestResponseText(after entryCount: Int) -> String {
+        for entry in dropFirst(entryCount).reversed() {
+            switch entry {
+            case .response:
+                return entry.textContent() ?? ""
+            case .prompt:
+                return ""
+            default:
+                continue
+            }
+        }
+        return ""
+    }
+}
